split files and create traits for composing logger
