+++ README.md
@@ -1,0 +1,8 @@
+
+
+## Notes
+
+- We implement against Logback
+- Can log to different formats (text, JSON, YAML, XML, ...)
+- Can accept maps, case classes, keys and values, ...
+- It is a Scala only library
