+++ src/main/scala/com/unstablebuild/slime/Format.scala
@@ -1,0 +1,7 @@
+package com.unstablebuild.slime
+
+trait Format {
+
+  def format(values: Seq[(String, Value)]): Array[Byte]
+
+}
