+++ src/main/scala/com/unstablebuild/slime/Logging.scala
@@ -1,0 +1,19 @@
+package com.unstablebuild.slime
+
+trait Logging {
+
+  protected def logger: Logger
+
+}
+
+trait LazyLogging extends Logging {
+
+  protected lazy val logger: Logger = Logger(getClass.getName)
+
+}
+
+trait StrictLogging extends Logging {
+
+  protected val logger: Logger = Logger(getClass.getName)
+
+}
