+++ src/main/scala/com/unstablebuild/slime/gen/MacroLogger.scala
@@ -1,0 +1,6 @@
+package com.unstablebuild.slime.gen
+
+import com.unstablebuild.slime.{Logger, SlimeLogger}
+
+@SlimeLogger
+class MacroLogger extends Logger {}
